command: Refuse to apply plan if its state doesn't match current
After running "terraform plan -out=tfplan" and then
"terraform apply tfplan" the plan file is left on disk and could
previously potentially be applied a second time.

Here we add a new constraint that prevents the use of a plan that was
not produced from the current state, thus avoiding that problem.

It will also reduce race conditions (on the human timescale) between
running "plan" and later running "apply", in environments where multiple
people/processes are using Terraform with the same remote state. This
hazard cannot be eliminated entirely without proper locking, but the
with this change in place the race condition is only for two concurrent
*applies*, as opposed to overlapping of the whole time period between plan
and apply.